fix: fix get countries info function
