+++ src/index.html
@@ -8,18 +8,26 @@
 </head>
 
 <body>
-    <section class="module--map">
-        <div id="map" class="map">
-            <div id="legend"></div>
-        </div>
-        <div class="map-tabs">
-            <button class="button button-map--tab" data-tab-name="totalCases">Total cases</button>
-            <button class="button button-map--tab" data-tab-name="totalRecovered">Total recovered</button>
-            <button class="button button-map--tab" data-tab-name="totalDeaths">Total deaths</button>
-            <button class="button button-map--tab" data-tab-name="deathsToCases">Deaths to cases</button>
-            <button class="button button-map--tab" data-tab-name="recoveredToCases">Recovered to cases</button>
-        </div>
-    </section>
+    <main class="main">
+        <section class="module-list">
+        </section>
+        <section class="module-map">
+            <div id="map" class="map">
+                <div id="legend"></div>
+            </div>
+            <div class="map-tabs">
+                <button class="button button-map--tab" data-tab-name="totalCases">Total cases</button>
+                <button class="button button-map--tab" data-tab-name="totalRecovered">Total recovered</button>
+                <button class="button button-map--tab" data-tab-name="totalDeaths">Total deaths</button>
+                <button class="button button-map--tab" data-tab-name="deathsToCases">Deaths to cases</button>
+                <button class="button button-map--tab" data-tab-name="recoveredToCases">Recovered to cases</button>
+            </div>
+        </section>
+        <section class="module-table">
+        </section>
+        <section class="module-graph">
+        </section>
+    </main>
     <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDTXoHvwIv-Vy60tjnRMJKDtyIv8PheamM"></script>
 </body>
