+++ src/styles/_base.scss
@@ -1,2 +1,41 @@
 /* Base: contained within this file are all 
 your resets, variables, mixins, and any utility classes. */
+
+$color_primary:#2A3132;
+$color_secondary:#336B87;
+$color_third:#7a3321;
+$color_ligth:#FFFFFF;
+@mixin p {
+    font: {
+        size: 16px;
+        family: Georgia;
+        style: normal;
+        weight: normal;
+    }
+    line-height: 130%;
+}
+
+@mixin button {
+    margin: 5px 20px;
+    cursor: pointer;
+    transition: 0.3s;
+    border: 0px;
+    background-color: transparent;
+}
+
+@mixin button-hover {
+    transition: 0.3s;
+    transform: scale(0.9);
+}
+
+@mixin button-active {
+    transition: 0.3s;
+    transform: scale(0.9);
+    border: 0px;
+}
+
+@mixin button__img {
+    width: 20px;
+    height: 20px;
+    page-break-after: 5px;
+}
