+++ src/styles/_layout.scss
@@ -18,3 +18,134 @@
     margin: 10px;
     border: 3px solid #000;
 }
+
+body {
+    background-color: #cccccc;
+    color: $color_ligth;
+}
+
+.main {
+    background-color: #5e5e5e;
+    max-width: 1200px;
+    margin: 0 auto;
+    display: flex;
+    flex-wrap: wrap;
+}
+
+.module-table {
+    min-width: 50%;
+    height: 300px;
+    background-color: $color_primary;
+    &__table {
+        margin: 10px auto;
+        width: 80%;
+        background-color: $color_third;
+        border: 1px solid $color_ligth;
+        &-thead {
+            background-color: $color_third;
+            &__figure {
+                margin: 5px auto;
+                &__icon {
+                    max-width: 40px;
+                }
+            }
+            &__search {
+                max-width: 140px;
+            }
+            &__search-tabs-container {
+                display: flex;
+                justify-content: center;
+                align-items: center;
+                &__tab {
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                }
+            }
+            &__names-columns,
+            &__target-info {
+                display: flex;
+                text-align: center;
+                font-weight: bold;
+                &__td-location,
+                &__td-infected,
+                &__td-recovered,
+                &__td-deaths {
+                    display: flex;
+                    flex-direction: column;
+                    justify-content: center;
+                    align-items: center;
+                    text-align: center;
+                    width: 25%;
+                }
+                img {
+                    width: 25px;
+                    height: 25px;
+                    padding: 5px;
+                    border-radius: 50%;
+                    border: 3px solid $color_primary;
+                    background-color: $color_ligth;
+                }
+            }
+        }
+        &-tbody {
+            background-color: $color_secondary;
+            display: block;
+            height: 120px;
+            overflow: auto;
+            &__tr {
+                display: flex;
+                text-align: right;
+                cursor: pointer;
+                &__td-location,
+                &__td-infected,
+                &__td-recovered,
+                &__td-deaths {
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                    text-align: center;
+                    width: 25%;
+                    padding: 2px;
+                }
+            }
+            &__tr:hover {
+                background-color: $color_third;
+            }
+        }
+    }
+    &__text-categories {
+        width: 100px;
+        text-align: center;
+    }
+    &__button {
+        @include button;
+    }
+    &__button {
+        @include button;
+        &__img {
+            @include button__img;
+        }
+    }
+    &__button:active {
+        @include button-active;
+    }
+}
+
+.module-list {
+    min-width: 50%;
+    height: 300px;
+    background-color: $color_secondary;
+}
+
+.module-map {
+    min-width: 50%;
+    height: 300px;
+    background-color: $color_third;
+}
+
+.module-graph {
+    min-width: 50%;
+    height: 300px;
+    background-color: $color_ligth;
+}
